backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
